Move activity list to features
