+++ src/features/activityOverview/api/getMonthNorm/handler.ts
@@ -1,8 +1,8 @@
 import { metadata } from './contract';
-import { RegisterHandler } from '../../server/shared';
-import { DayType } from '../../server/workHoursNorm';
-import { addDays, getRange, getMonthBoundaries } from '../../shared/date';
-import { assertNever } from '../../shared';
+import { RegisterHandler } from 'server/shared';
+import { DayType } from 'server/workHoursNorm';
+import { addDays, getRange, getMonthBoundaries } from 'shared/date';
+import { assertNever } from 'shared';
 
 const handler: RegisterHandler = (server, { calendar }) => server.route({
     ...metadata,
