+++ src/features/activityOverview/api/getMonthNorm/index.ts
@@ -1,5 +1,5 @@
 import { metadata, GetMonthNormResponse } from './contract';
-import { ensureSuccessStatusCode, readJson } from '../../shared/http';
+import { ensureSuccessStatusCode, readJson } from 'shared/http';
 
 export const getMonthNorm = async (): Promise<GetMonthNormResponse> => {
     const response = await fetch(metadata.url, { method: metadata.method });
