+++ src/features/activityOverview/api/queryTimings/index.ts
@@ -1,6 +1,6 @@
 import { metadata, QueryTimingsRequest, QueryTimingResponse } from './contract';
-import { ensureSuccessStatusCode, readJson, formatUrl } from '../../shared/http';
-import { TimesheetEntry } from '../../shared/types';
+import { ensureSuccessStatusCode, readJson, formatUrl } from 'shared/http';
+import { TimesheetEntry } from 'shared/types';
 
 export const queryTimings = async (req: QueryTimingsRequest): Promise<QueryTimingResponse> => {
     const query = {
