+++ src/features/activityOverview/components/activityHeatmap/index.scss
