+++ src/features/activityOverview/components/activityList/header.scss
