+++ src/features/activityOverview/components/activityList/index.scss
