+++ src/features/activityOverview/components/hoursGauge/index.scss
