+++ src/features/activityOverview/components/hoursGauge/index.tsx
