+++ src/features/activityOverview/state/activityList/reducer.ts
@@ -1,4 +1,4 @@
-import { assertNever } from '../../shared';
+import { assertNever } from 'shared';
 import { ActivityListAction, ActivityListState } from './types';
